feat: improve handling of "abbreviated" flake refs (#26)
If not well defined, i.e. if the flakeref cannot be parsed as a url,
we resolve it either as an indirect flake or local path.

Note: if not "well-defined" parsing flakerefs is "impure",
      i.e. depends on the state of the local system (files).
      The resulting flakeref however, serializes into well-defined form.


Resolving an abbreviated URL to local files:

Nix supports referring to local files/paths without an explicit scheme.
However, it distinguishes between flakes in git repositories
(disambiguated using `git+file://`) from "free" ones (`path:`).
The logic for this in nix can be found at
<https://github.com/NixOS/nix/blob/bf7dc3c7dc24f75fa623135750e8f10b8bcd94f9/src/libexpr/flake/flakeref.cc#L112C5-L197>
This method implements a subset of this logic assuming we require the
path to exist and to be a flake.
In abstract we
1. locate a flake (by finding a `flake.nix` in all ancestors)
  - fail if directory does not exist
  - fail if skipping file system boundaries
  - fail if "exiting" a git repository (if there is one)
2. locate the root of the "containing" git repo (if applicable)
  - check whether `.git/shallow` exists and track in the url params
  - ensure there is no `?dir=` param if flake is in a subdir
3. construct a `file+git:` or `path:` url as required


Additionally this PR refactors the parsing of flakerefs in
https://github.com/flox/runix/commit/5c91d2a4f3b2752d6ecac6127311cad2a3d3f4c8
to be URL first.

So far, all flakeref parsers implemented the parsing as `FromStr` only.
In all cases `FromStr` would first parse the input as `Url` and then
process mainly the Url.
After this refactoring all flakerefs can only be _purely_ parsed using
Urls or "well-defined" url strings (see above).
`FromStr` is now uniformly implemented as

    fn from_str(s: &str) -> Result<Self, Self::Err> {
        let url = Url::parse(s)?;
        Self::from_url(url)
    }

for all `FlakeRef` types.

If "impure" resolution is necessary, `FlakeRef::resolve_local` provides
an interface to create a well-defined Url.
`FlakeRef::from_str` by always tries to impurely resolve urls for inputs
that are not well-defined.

The change makes `FlakeRef` instances pure by default, which simplifies
the parsing as impure components are stripped out.