feat: wrap Nix URI parser (#38)
This PR adds the `url_parser` module, which provides a wrapper around
the `flox/parser-util` binary, which is itself an interface to Nix's own
URI parsing routines.

The mechanism for parsing and usage inside `runix` looks like this:
- The `parser-util` binary exists somewhere on the system
- To resolve the URI, one of `url_parser::{resolve_flake_ref,
lock_flake_ref, installables_flake_ref} is called
- This Rust function calls `parser-util (-r | -l | -i) <flake ref>`,
which returns JSON
- A faithful representation of the JSON is constructed, regardless of
its usage elsewhere in `runix` i.e. all fields in the JSON are parsed
and retained regardless of whether we actually use them. This is done so
that we can use those fields at a later date without needing to change
the URI parser wrapper.
- Parsing is done in two stages, the first being a fully generic parsing
to a typed representation of a union of the fields that could appear in
any of the `-r`, `-l`, or `-i` outputs from `parser-util`. The second
phase tries to fit this into the particular output that was requested
(resolved, locked, etc).
- This is necessary because if URI resolution fails `parser-util` will
still return some information, just not all of it e.g. the `resolvedRef`
attribute. We want to distinguish between a legitimate failure to parse
the JSON output (wrong type, can't convert data to a certain type, etc)
and a resolution failure.
- Once a typed, parsed flake reference is constructed (one of
`ResolvedFlakeRef`, `LockedFlakeRef`, or `InstallableFlakeRef`) you can
then fit it into `runix::flake_ref::FlakeRef` via the `TryFrom` trait.
- Currently URIs are resolved with `resolve_flake_ref` and `FlakeRef`s
are constructed from the `originalRef` JSON attributes, but changing
this to `resolvedRef` would be trivial (you'd have to fix tests that
expect `originalRef` attributes).

The path to the `parser-util` binary is now embedded at compile-time via `url_parser::PARSER_UTIL_BIN_PATH`, and this variable should be used to path the `parser-util` path to functions that need it.

---------

Co-authored-by: Yannik Sander <yannik@floxdev.com>