fix: use runix parser instead of parser-util
parser-util does not yet accept Nix config, which means that it cannot
yet be used by flox, so temporarily use the runix parser until
parser-util is ready for use by flox.

In more depth, parser-util uses the flake registry to resolve a
FlakeRef. flox augments the flake registry with Nix config, but it can't
currently pass that config to parser-util. So FlakeRef::from_str fails
when parsing indirect refs that are in the flox flake registry.